refactor: split LayoutConverter by responsibility
